requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
Babel==2.5.3
bcrypt==3.1.4
blinker==1.4
boto3==1.8.5
botocore==1.11.5
cffi==1.13.2
click==6.7
docutils==0.14
Flask==0.12.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.1.0
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
s3transfer==0.1.13
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.7
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.1
